+++ src/gort/overwatcher/observer.py
@@ -130,6 +130,8 @@ class ObserverOverwatcher(OverwatcherModule):
         self._starting_observations: bool = False
         self._cancelling: bool = False
 
+        self.force_focus: bool = False  # Force focus before the next tile
+
     @property
     def is_observing(self) -> bool:
         """Returns whether the observer is currently observing."""
@@ -232,7 +234,6 @@ class ObserverOverwatcher(OverwatcherModule):
         observer = self.gort.observer
 
         n_tile_positions = 0
-        force_focus: bool = False
 
         while True:
             exp: Exposure | bool = False
@@ -250,7 +251,7 @@ class ObserverOverwatcher(OverwatcherModule):
                     f"observing dither positions {tile.dither_positions}."
                 )
 
-                await self.check_focus(force=force_focus or n_tile_positions == 0)
+                await self.check_focus(force=n_tile_positions == 0 or self.force_focus)
 
                 for dpos in tile.dither_positions:
                     await self.overwatcher.troubleshooter.wait_until_ready(300)
@@ -312,7 +313,7 @@ class ObserverOverwatcher(OverwatcherModule):
                         break
                     if self.focusing:
                         # Force a new focus if the error occurred while focusing.
-                        force_focus = True
+                        self.force_focus = True
 
                     continue
 
@@ -378,6 +379,7 @@ class ObserverOverwatcher(OverwatcherModule):
                 raise
             else:
                 self.focusing = False
+                self.force_focus = False
 
     async def pre_observe_checks(self):
         """Runs pre-observe checks."""